refactor(context): POC decoupling of response creation from context class
split creating Responses to its own class, and let Context class extend the response creator class, this allows the current header objects init optimizations to remain unchanged
